just using await, structuring validate with Joi

+++ controllers/product.controller.js
@@ -1,28 +1,23 @@
 const product = require('../model/product.model');
 const user = require('../model/user.model');
 const Joi = require('joi');
-const { checkProductSchema } = require('./checkInput');
+const { checkProductSchema } = require('./productValidate');
 const jwt = require('jsonwebtoken');
 
 
-module.exports.getProduct = function(req, res) {
-    product.find({ status: "active" }, function(error, result) {
-        if (error) res.json(error);
-        else res.json(result);
-    })
+module.exports.getProduct = async function(req, res) {
+    const result = await product.find({ status: "active" });
+    res.json(result);
 }
 
 module.exports.getMyProduct = async function(req, res) {
     try {
         const authHeader = req.headers['authorization'];
+        if (!authHeader) throw new Error('You are not login');
         const token = authHeader.split(' ')[1];
         const checkUser = await jwt.verify(token, process.env.secret_key);
-        await product.find({ owner: checkUser.email }, (err, result) => {
-            if (err) throw new Error(err.message);
-            else res.json({
-                result
-            })
-        })
+        const result = await product.find({ owner: checkUser.email });
+        res.json(result);
     } catch (err) {
         res.json({
             message: err.message
@@ -40,31 +35,12 @@ module.exports.changeStatusProduct = async function(req, res) {
 
         const check = await product.findOne({ _id: productId, owner: owner.email });
         if (!check) throw new Error('You are not allowed to access');
-
-        await product.findOne({ _id: productId, owner: owner.email }, async(err, result) => {
-            if (err) throw new Error(err.message)
-            else {
-                if (result.status == 'active') {
-                    await product.findOneAndUpdate({ _id: result._id }, { $set: { status: 'hide' } }, (err, result) => {
-                        if (err) throw new Error(err.message);
-                    })
-                } else await product.findOneAndUpdate({ _id: result._id }, { $set: { status: 'active' } })
-                res.json({
-                    code: 0,
-                    message: " Change status successful",
-                    data: result
-                })
-            }
+        const result = await product.findOneAndUpdate({ _id: productId, owner: owner.email }, { $set: { status: req.body.status } });
+        res.json({
+            code: 0,
+            message: 'Change status successful',
+            data: result
         })
-
-        // await product.findOneAndUpdate({ _id: productId, owner: owner.email }, { status: 'hide' }, (err, result) => {
-        //     if (err) throw new Error(err.message)
-        //     else res.json({
-        //         code: 0,
-        //         message: "Hide product successful",
-        //         data: result
-        //     })
-        // })
     } catch (err) {
         res.json({
             message: err.message
@@ -84,13 +60,11 @@ module.exports.updateProduct = async function(req, res) {
         const check = await product.findOne({ _id: productId, owner: owner.email });
         if (!check) throw new Error('You are not allowed to access');
 
-        await product.findOneAndUpdate({ _id: productId, owner: owner.email }, { $set: req.body }, (err, result) => {
-            if (err) throw new Error(err.message)
-            else res.json({
-                code: 0,
-                message: " Update successful",
-                data: result
-            })
+        await product.findOneAndUpdate({ _id: productId, owner: owner.email }, { $set: req.body });
+        res.json({
+            code: 0,
+            message: " Update successful",
+            data: result
         })
     } catch (err) {
         res.json({
@@ -109,17 +83,12 @@ module.exports.removeProduct = async function(req, res) {
         const check = await product.findOne({ _id: productId, owner: owner.email });
         if (!check) throw new Error('You are not allowed to access');
 
-        await product.findOneAndDelete({
-            _id: productId,
-            owner: owner.email
-        }, (err, result) => {
-            if (err) throw new Error(err.message)
-            else res.json({
-                code: 0,
-                message: 'Delete successful',
-                data: result
-            })
-        });
+        await product.findOneAndDelete({ _id: productId, owner: owner.email });
+        res.json({
+            code: 0,
+            message: 'Delete successful',
+            data: result
+        })
 
     } catch (err) {
         res.json({
@@ -132,24 +101,19 @@ module.exports.removeProduct = async function(req, res) {
 module.exports.postProduct = async function(req, res) {
 
     try {
-
-        Joi.validate(req.body, checkProductSchema, (err, result) => {
-            if (err) throw new Error(err.message);
-        })
+        await Joi.validate(req.body, checkProductSchema);
         if (!req.body.name) {
             throw new Error('Name of product is required');
         }
         if (!req.body.price) {
             throw new Error('Price of product is required');
         }
-
-        let newProduct = await product.create({
+        const newProduct = await product.create({
             owner: req.user.email,
             name: req.body.name,
             price: req.body.price,
             status: 'active'
         });
-
         res.json({
             code: 0,
             data: newProduct,
